requirements.txt
pyocclient==0.6
python-dotenv==1.0.1
requests==2.32.3

